+++ client/src/components/TopCatPage/CategoryCard/categoryCard.css
@@ -1,16 +1,5 @@
-#item3 {
-  background-image: url("https://cdn3.yoox.biz/cloud/balenciagawp/uploads/2019/10/19.jpg");
-  /* background-image: url("https://www.balenciaga.com/66/41/41936513ht_14_a_g.jpg"); */
-}
-#item4 {
-  background-image: url("https://cdn3.yoox.biz/cloud/balenciagawp/uploads/2019/10/69.jpg");
-}
-#item1 {
-  background-image: url("https://www.balenciaga.com/66/41/41922379IU_12_a_g.jpg");
-}
-#item2 {
-  /* background-image: url("https://www.balenciaga.com/66/41/41809866SC_12_a_g.jpg"); */
-  background-image: url("https://cdn3.yoox.biz/cloud/balenciagawp/uploads/2019/10/56.jpg");
+.menuItem :hover {
+  background-color: black;
 }
 
 .hidden-link {
@@ -21,14 +10,17 @@
 }
 
 .top-cat-name {
+  display: block;
   font-size: 42px;
   background: white;
   width: 100%;
   text-align: center;
+  text-transform: uppercase;
 }
 
 .category-item-home {
   background-size: cover;
+  /* background-image: url("https://www.balenciaga.com/66/41/41809866SC_12_a_g.jpg"); */
   background-position: center;
   filter: grayscale(1);
   display: flex;
@@ -37,7 +29,7 @@
   align-items: center;
   font-size: 40px;
   text-decoration: none;
-  transition: all 1s ease-in;
+  transition: all 2s ease;
   height: 75vh;
 }
