+++ client/src/components/TopCatPage/CategoryCard/index.js
@@ -1,19 +1,26 @@
-import React from 'react';
-import {Link} from "react-router-dom";
+import React from "react";
+import { Link, Grid } from "@material-ui/core";
+import { ThemeProvider } from "@material-ui/core/styles";
+import theme from "../../../theme";
+import useStyles from "./useStyles";
+import "./categoryCard.css";
 
+function CategoriesMenuItem(props) {
+  const { name, image } = props;
+  const classes = useStyles();
 
-export default props => {
-    const {item} = props;
+  return (
+    // <ThemeProvider theme={theme}>
+    <Grid
+      className="category-item-home"
+      item
+      xs={3}
+      style={{ backgroundImage: `url(${image})` }}
+    >
+      <p className="top-cat-name">{name}</p>
+    </Grid>
+    // </ThemeProvider>/
+  );
+}
 
-    return (
-        <>
-            <div
-                className="caterory-gallery-item col-12 col-md-6 col-lg-4 col-xl-2 card-body p-5 border border-secondary text-center"
-            >
-            <Link to={`/categories/${item.name}`} className ="p-5 text-center d-block">
-                <p>{props.item.name}</p>
-            </Link>
-            </div>
-        </>
-    )
-};
+export default CategoriesMenuItem;
